+++ src/main/java/beauty/scheduler/dao/MappersStorage.java
@@ -12,7 +12,7 @@ import beauty.scheduler.entity.enums.ServiceType;
 //right now extracted most such things to one place
 public interface MappersStorage {
 
-    EntityMapper<User> USER_ENTITY_MAPPER = resultSet -> new User(resultSet.getLong("users_id"),
+    EntityMapper<User> USER_ENTITY_MAPPER = resultSet -> new User(resultSet.getInt("users_id"),
             resultSet.getString("users_first_name_en"),
             resultSet.getString("users_last_name_en"),
             resultSet.getString("users_first_name_uk"),
@@ -23,7 +23,7 @@ public interface MappersStorage {
             Role.lookupNotNull(resultSet.getString("users_role")),
             ServiceType.lookupNotNull(resultSet.getString("users_service_type")));
 
-    EntityMapper<EmailMessage> EMAIL_MESSAGE_ENTITY_MAPPER = resultSet -> new EmailMessage(resultSet.getLong("email_messages_id"),
+    EntityMapper<EmailMessage> EMAIL_MESSAGE_ENTITY_MAPPER = resultSet -> new EmailMessage(resultSet.getInt("email_messages_id"),
             resultSet.getString("email_messages_email"),
             resultSet.getString("email_messages_subject"),
             resultSet.getString("email_messages_text_message"),
@@ -31,11 +31,11 @@ public interface MappersStorage {
             resultSet.getBoolean("email_messages_sent"),
             resultSet.getString("email_messages_quick_access_code"));
 
-    EntityMapper<Appointment> APPOINTMENT_ENTITY_MAPPER = resultSet -> new Appointment(resultSet.getLong("appointments_id"),
+    EntityMapper<Appointment> APPOINTMENT_ENTITY_MAPPER = resultSet -> new Appointment(resultSet.getInt("appointments_id"),
             resultSet.getDate("appointments_appointment_date").toLocalDate(),
             resultSet.getByte("appointments_appointment_time"),
             ServiceType.lookupNotNull(resultSet.getString("appointments_service_type")),
-            new User(resultSet.getLong("customer_id"),
+            new User(resultSet.getInt("customer_id"),
                     resultSet.getString("customer_first_name_en"),
                     resultSet.getString("customer_last_name_en"),
                     resultSet.getString("customer_first_name_uk"),
@@ -45,7 +45,7 @@ public interface MappersStorage {
                     resultSet.getString("customer_password"),
                     Role.lookupNotNull(resultSet.getString("customer_role")),
                     ServiceType.lookupNotNull(resultSet.getString("customer_service_type"))),
-            new User(resultSet.getLong("master_id"),
+            new User(resultSet.getInt("master_id"),
                     resultSet.getString("master_first_name_en"),
                     resultSet.getString("master_last_name_en"),
                     resultSet.getString("master_first_name_uk"),
@@ -57,12 +57,12 @@ public interface MappersStorage {
                     ServiceType.lookupNotNull(resultSet.getString("master_service_type"))),
             resultSet.getBoolean("appointments_service_provided"));
 
-    EntityMapper<Feedback> FEEDBACK_ENTITY_MAPPER = resultSet -> new Feedback(resultSet.getLong("feedbacks_id"),
-            new Appointment(resultSet.getLong("appointment_id"),
+    EntityMapper<Feedback> FEEDBACK_ENTITY_MAPPER = resultSet -> new Feedback(resultSet.getInt("feedbacks_id"),
+            new Appointment(resultSet.getInt("appointment_id"),
                     resultSet.getDate("appointment_appointment_date").toLocalDate(),
                     resultSet.getByte("appointment_appointment_time"),
                     ServiceType.lookupNotNull(resultSet.getString("appointment_service_type")),
-                    new User(resultSet.getLong("customer_id"),
+                    new User(resultSet.getInt("customer_id"),
                             resultSet.getString("customer_first_name_en"),
                             resultSet.getString("customer_last_name_en"),
                             resultSet.getString("customer_first_name_uk"),
@@ -72,7 +72,7 @@ public interface MappersStorage {
                             resultSet.getString("customer_password"),
                             Role.lookupNotNull(resultSet.getString("customer_role")),
                             ServiceType.lookupNotNull(resultSet.getString("customer_service_type"))),
-                    new User(resultSet.getLong("master_id"),
+                    new User(resultSet.getInt("master_id"),
                             resultSet.getString("master_first_name_en"),
                             resultSet.getString("master_last_name_en"),
                             resultSet.getString("master_first_name_uk"),
